refactor(通讯配置): 重构通讯方式和灵巧手型号配置
将 USE_CANFD 宏重命名为 COMMUNICATION_MODE 以提升可读性
添加 HAND_TYPE 宏用于配置灵巧手型号
更新 README 文档以反映配置变更
修复静态类型转换警告

+++ src/lhandpro_service/src/CANFDMaster.cpp
@@ -72,7 +72,7 @@ bool CANFDMaster::connect(int deviceIndex, int nomBaud, int dataBaud) {
 
   // 扫描可用的CAN接口
   std::vector<std::string> devices = scanDevices();
-  if (deviceIndex < 0 || deviceIndex >= devices.size()) {
+  if (deviceIndex < 0 || deviceIndex >= static_cast<int>(devices.size())) {
     return false;
   }
